--- a/git.h
+++ b/git.h
@@ -35,7 +35,7 @@ extern void git (const git_uint8 * game,
                  git_uint32 cacheSize,
                  git_uint32 undoSize);
 
-extern void fatalError (const char *);
+extern noreturn void fatalError (const char *);
 
 // memory.c