--- a/savefile.c
+++ b/savefile.c
@@ -80,7 +80,7 @@ git_sint32 restoreFromFile (git_sint32 * base, git_sint32 id,
             for (i = 0 ; i < 128 ; ++i)
             {
                 glui32 c = glk_get_char_stream (file);
-                if (gRom [i] != c)
+                if (gInitMem [i] != c)
                     return 1;
             }
         }
@@ -125,7 +125,7 @@ git_sint32 restoreFromFile (git_sint32 * base, git_sint32 id,
                 
                 for (++mult ; mult > 0 ; --mult, ++i)
                     if (i >= protectEnd || i < protectPos)
-                        gRam [i] = gRom [i] ^ c;
+                        gMem [i] = gInitMem [i] ^ c;
             }
 
             while (i < gEndMem && bytesRead < chunkSize)
@@ -142,16 +142,16 @@ git_sint32 restoreFromFile (git_sint32 * base, git_sint32 id,
                 
                 for (++mult ; mult > 0 ; --mult, ++i)
                     if (i >= protectEnd || i < protectPos)
-                        gRam [i] = c;
+                        gMem [i] = c;
             }
 
             while (i < gExtStart)
                 if (i >= protectEnd || i < protectPos)
-                    gRam [i] = gRom [i], ++i;
+                    gMem [i] = gInitMem [i], ++i;
 
             while (i < gEndMem)
                 if (i >= protectEnd || i < protectPos)
-                    gRam [i] = 0, ++i;
+                    gMem [i] = 0, ++i;
 
             if (bytesRead != chunkSize)
                 return 1; // Too much data!
@@ -245,7 +245,7 @@ git_sint32 saveToFile (git_sint32 * base, git_sint32 * sp, git_sint32 id)
     // Header chunk.
     glk_put_string ("IFhd");
     writeWord (128);
-    glk_put_buffer ((char *) gRom, 128);
+    glk_put_buffer ((char *) gInitMem, 128);
 
     // Stack chunk.
     glk_put_string ("Stks");
@@ -271,8 +271,8 @@ git_sint32 saveToFile (git_sint32 * base, git_sint32 * sp, git_sint32 id)
     writeWord (gEndMem);
     for (zeroCount = 0, n = gRamStart ; n < gEndMem ; ++n)
     {
-        unsigned char romC = (n < gExtStart) ? gRom[n] : 0;
-        unsigned char c = ((git_uint32) romC) ^ ((git_uint32) gRam[n]);
+        unsigned char romC = (n < gExtStart) ? gInitMem[n] : 0;
+        unsigned char c = ((git_uint32) romC) ^ ((git_uint32) gMem[n]);
         if (c == 0)
             ++zeroCount;
         else