So I've been looking at the Git source code and found a couple of things that could be improved. For a scripted play-through of Counterfeit Monkey release 4 I get the following results on my laptop (Intel Core 2 Duo in 64 bit mode on Linux):
* gcc-4.9 -O2 -DUSE_INLINE: from 239s to 190s (79%)
* clang-3.4 -O2 -DUSE_INLINE: from 304s to 224s (74%)

The first few patches actually just fix various things that GCC and/or clang complain about. The optimization changes follow. None of the patches make large changes to the code, they mostly nibble away at small inefficiencies at the edges. The core of the compiler/interpreter is rather inscrutable :)