@DavidKinder 2015-06-13
HI Peter, I have finally got round to looking at these, and I've included all your patches, except for "Optimized memory access code." I wasn't entirely happy about this one as I am unsure as to what GCC version introduced the byteswap intrinsic, and I find that older GCCs still give much better performance on straight C code - on my system git compiled with GCC 2.95 is about 30% faster than when compiled with GCC 4.8.